requirements.txt
osxphotos
nicegui
pillow
pillow-heif